Move png encoding into common
